refactor: Move errors to own module & file
