+++ src/lib.rs
@@ -6,6 +6,7 @@ pub mod callback_handler;
 pub mod callback_requests;
 pub mod cluster_context;
 pub mod constants;
+pub mod errors;
 pub(crate) mod policy;
 pub mod policy_artifacthub;
 pub mod policy_evaluator;
